Move code to prepare for CoreGroup feauture flag (#640)
