+++ openmls/src/framing/mod.rs
@@ -26,10 +26,20 @@ pub use errors::*;
 pub use message::*;
 pub use plaintext::*;
 pub use sender::*;
+use tls_codec::{TlsDeserialize, TlsSerialize, TlsSize};
 pub use validation::*;
 
 #[cfg(test)]
 mod test_framing;
+
+#[derive(
+    PartialEq, Clone, Copy, Debug, Serialize, Deserialize, TlsDeserialize, TlsSerialize, TlsSize,
+)]
+#[repr(u8)]
+pub enum WireFormat {
+    MlsPlaintext = 1,
+    MlsCiphertext = 2,
+}
 
 /// This struct is used to group common framing parameters
 /// in order to reduce the number of arguments in function calls.
