+++ openmls/src/group/core_group/test_proposals.rs
@@ -8,12 +8,12 @@ use crate::{
     credentials::{CredentialBundle, CredentialType},
     extensions::{Extension, ExtensionType, KeyIdExtension, RequiredCapabilitiesExtension},
     framing::sender::{Sender, SenderType},
-    framing::{FramingParameters, MlsPlaintext},
+    framing::{FramingParameters, MlsPlaintext, WireFormat},
     group::{
         create_commit_params::CreateCommitParams,
         errors::CoreGroupError,
         proposals::{CreationProposalQueue, ProposalStore, StagedProposal, StagedProposalQueue},
-        GroupContext, GroupEpoch, GroupId, WireFormat,
+        GroupContext, GroupEpoch, GroupId,
     },
     key_packages::{KeyPackageBundle, KeyPackageError},
     messages::proposals::{AddProposal, Proposal, ProposalOrRef, ProposalReference, ProposalType},
