+++ openmls/src/group/mod.rs
@@ -89,57 +89,3 @@ impl PartialOrd for GroupEpoch {
         self.0.partial_cmp(&other.0)
     }
 }
-
-#[derive(
-    Debug, Clone, PartialEq, Serialize, Deserialize, TlsSerialize, TlsDeserialize, TlsSize,
-)]
-pub struct GroupContext {
-    group_id: GroupId,
-    epoch: GroupEpoch,
-    tree_hash: TlsByteVecU8,
-    confirmed_transcript_hash: TlsByteVecU8,
-    extensions: TlsVecU32<Extension>,
-}
-
-#[cfg(any(feature = "test-utils", test))]
-impl GroupContext {
-    pub(crate) fn set_epoch(&mut self, epoch: GroupEpoch) {
-        self.epoch = epoch;
-    }
-}
-
-/// Configuration for core group.
-#[derive(Clone, Copy, Debug)]
-pub struct CoreGroupConfig {
-    /// Flag whether to send the ratchet tree along with the `GroupInfo` or not.
-    /// Defaults to false.
-    pub add_ratchet_tree_extension: bool,
-    pub padding_block_size: u32,
-    pub additional_as_epochs: u32,
-}
-
-impl CoreGroupConfig {
-    /// Get the padding block size used in this config.
-    pub fn padding_block_size(&self) -> u32 {
-        self.padding_block_size
-    }
-}
-
-impl Default for CoreGroupConfig {
-    fn default() -> Self {
-        Self {
-            add_ratchet_tree_extension: false,
-            padding_block_size: 10,
-            additional_as_epochs: 0,
-        }
-    }
-}
-
-#[derive(
-    PartialEq, Clone, Copy, Debug, Serialize, Deserialize, TlsDeserialize, TlsSerialize, TlsSize,
-)]
-#[repr(u8)]
-pub enum WireFormat {
-    MlsPlaintext = 1,
-    MlsCiphertext = 2,
-}
